Fix battle intent bugs
# Abstract
Fixed bugs in the battle intent.  Mainly this involved adding input validation. Added code to check for "value" slot presence before dereferencing.  If there are input errors, then respond with a notice that we couldn't understand the input.

For sample utterances, I removed half the utterances for Dice Logic because the Amazon skill processer simply couldn't figure out how to set `numDice` properly if the word "me" wasn't in the input.  So I removed the sample utterances without the word "me".  This is a bit drastic, but if we can figure out why this doesn't work, we can add back those other sample utterances.

# Testing
None other than the removal of the sample utterances.